\name{carleton-package}
\alias{carleton-package}
\alias{carleton}
\docType{package}
\title{
Functions for introductory statistics classes at Carleton College
}
\description{
Custom functions written for students taking statistics at
Carleton College. Functions for the bootstrap and
permutation tests as well as a grouped barplot. There are also some
demonstrations to illustrate the concepts of correlation, the normal
quantile-quantile plot and confidence intervals.
}
\details{
\tabular{ll}{
Package: \tab carleton\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2013-10-22\cr
License: GPL-2
}

}
\author{
Laura Chihara

Maintainer: Laura Chihara <lchihara@carleton.edu>

}
