\name{Milkshakes}
\alias{Milkshakes}
\docType{data}
\title{
Milkshakes (chocolate)
}
\description{
Nutrional information on chocolate milkshakes from a sample of restaurants.

}
\usage{data(Milkshakes)}
\format{
  A data frame with 29 observations on the following 11 variables.
  \describe{
    \item{\code{Restaurant}}{Names of restaurants}
    \item{\code{Type}}{Type of restaurant, \code{Dine In} \code{Fast Food}}
    \item{\code{Calories}}{Calories per serving}
    \item{\code{Fat}}{Fat per serving (g)}
    \item{\code{Sodium}}{Sodium per serving (mg)}
    \item{\code{Carbs}}{Carbohydrates per serving (g)}
    \item{\code{SizeOunces}}{Size of milkshake (ounces)}
    \item{\code{CalPerOunce}}{Calories per ounce}
    \item{\code{FatPerOunce}}{Fat per ounce}
    \item{\code{CarbsPerOunce}}{Carbohydrates per ounce}
  }
}
\details{
Nutritional information on chocolate milkshakes from a sample of restaurants.
}
\source{
Data collected by Carleton students Yoni Blumberg (2013) and Lindsay Guthrie (2013).
}

\keyword{datasets}
