\name{states03}
\alias{states03}
\docType{data}
\title{
US States data, 2003
}
\description{
Census data on the 50 states from 2003.
}
\usage{data(states03)}
\format{
  A data frame with 50 observations on the following 24 variables.
  \describe{
    \item{\code{State}}{the 50 states}
    \item{\code{Region}}{a factor with levels \code{Midwest}, \code{Northeast}, \code{South}, \code{West}}
    \item{\code{Pop}}{Population in 1000}
    \item{\code{Births}}{Number of births}
    \item{\code{Deaths}}{Number of deaths}
    \item{\code{Pop18}}{Percent of population 18 years of age or younger}
    \item{\code{Pop65}}{Percent of population 65 years of age or older}
    \item{\code{HSGrad}}{Percent of population 25 years of age or older with a high school degree}
    \item{\code{ColGrad}}{Percent of population 25 yeras of age or older with a college degree}
    \item{\code{TeacherPay}}{Average teachers salary in dollars}
    \item{\code{InfMortality}}{Infant mortality per 1000 live births}
    \item{\code{TeenBirths}}{Live births per 1000 15-19 year old females}
    \item{\code{ViolentCrime}}{Violent crime per 100000 population}
    \item{\code{PropertyCrime}}{Property crime per 100000 population}
    \item{\code{DeathPenalty}}{State has death penalty?}
    \item{\code{Executions}}{Number of executions 1977-2003}
    \item{\code{Poverty}}{Percent of populaton below the poverty level}
    \item{\code{Unemp}}{Percent unemployed (of population 16 years or older)}
    \item{\code{Uninsured}}{Percent uninsured (3 year aveage)}
    \item{\code{Income}}{Median household income in 1998 dollars}
    \item{\code{Earnings}}{Average hourly earnings of production workers in
      manufacturing}
    \item{\code{Heart}}{Deaths by heart disease per 100000 population}  
   \item{\code{Vehicles}}{Deaths by motor vehicle accidents per 100000 population}
   \item{\code{Homeowners}}{Home ownership rate}   
  }
}
\details{
Data on the 50 states.
}
\source{
United States Census Bureau \url{http://www.census.gov/}
}
\keyword{datasets}
